Fix : homePage index -> home 으로 파일이동
