+++ src/pages/index.tsx
@@ -1,24 +1,12 @@
-import SideBar from 'components/Common/organisms/Sidebar'
 import { MainTemplates } from 'components/Common/templates/MainTemplates/style'
-import { TimeBoard, Profile, NoticeBoard, SelfStudyBoard, MassageBoard, MealBoard } from 'components/Home/organisms'
-import { HomeTemplates } from 'components/Home/templates/HomeTemplates/style'
 import UseThemeEffect from 'hooks/useThemeEffect'
 
-const HomePage = () => {
+const PromotionPage = () => {
   UseThemeEffect()
   return (
-    <MainTemplates>
-      <SideBar/>
-      <HomeTemplates>
-        <TimeBoard/>
-        <Profile/>
-        <MealBoard/>
-        <NoticeBoard/>
-        <SelfStudyBoard/>
-        <MassageBoard/>
-      </HomeTemplates>
-   </MainTemplates>
+    <>
+    </>
   )
 }
 
-export default HomePage
+export default PromotionPage
